Added win state

+++ Game/Game_Jam_Nov_2019/Assets/Scripts/WinState.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class WinState : MonoBehaviour
 {
-    public float score;
+    public int score;
     public Vector3 scoreCenter;
     public Vector3 scoreSize;
     // Start is called before the first frame update
